Move vehicle-related classes to own module
