+++ bimmer_connected/account.py
@@ -21,7 +21,7 @@ from bimmer_connected.api.authentication import MyBMWAuthentication
 from bimmer_connected.api.client import MyBMWClient, MyBMWClientConfiguration
 from bimmer_connected.api.regions import Regions
 from bimmer_connected.const import VEHICLES_URL, CarBrands
-from bimmer_connected.models import GPSPosition
+from bimmer_connected.vehicle.models import GPSPosition
 from bimmer_connected.vehicle import ConnectedDriveVehicle
 
 VALID_UNTIL_OFFSET = datetime.timedelta(seconds=10)
