+++ bimmer_connected/vehicle/__init__.py
@@ -1,0 +1,1 @@
+from .vehicle import DriveTrainType, VehicleViewDirection, ConnectedDriveVehicle
