+++ bimmer_connected/vehicle/models.py
