+++ bimmer_connected/vehicle/vehicle_status.py
